ported the kotlin code generator to use a base class tests pass
